# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageviewRequest": {
        "properties": {
          "api_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "referrer": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referrer"
          },
          "session_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "api_key",
          "session_id",
          "url"
        ],
        "title": "PageviewRequest",
        "type": "object"
      },
...
      "TapPoint": {
        "description": "One tap. ``x`` is a fraction of the document width; ``y`` is CSS px\nfrom the document top (viewport px when the target is fixed or sticky).",
        "properties": {
          "el": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "El"
          },
          "x": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "TapPoint",
        "type": "object"
      },
      "TapsRequest": {
        "description": "Body of ``POST /api/v1/taps``: the taps and scroll depth of one pageview.\n\n``session_id`` is accepted for parity with the other ingestion bodies but\nis never stored.",
        "properties": {
          "api_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "path": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "scroll": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scroll"
          },
          "session_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "taps": {
            "items": {
              "$ref": "#/components/schemas/TapPoint"
            },
            "maxItems": 50,
            "title": "Taps",
            "type": "array"
          },
          "viewport": {
            "enum": [
              "mobile",
              "tablet",
              "desktop"
            ],
            "title": "Viewport",
            "type": "string"
          },
          "vw": {
            "maximum": 10000.0,
            "minimum": 100.0,
            "title": "Vw",
            "type": "integer"
          }
        },
        "required": [
          "api_key",
          "session_id",
          "path",
          "viewport",
          "vw"
        ],
        "title": "TapsRequest",
        "type": "object"
      },
      "TrackEventRequest": {
        "properties": {
          "api_key": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Api Key",
            "type": "string"
          },
          "event_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Event Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "session_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          },
          "timestamp": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "required": [
          "api_key",
          "event_name",
          "session_id"
        ],
        "title": "TrackEventRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Internal-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted, privacy-first analytics platform controlled entirely through a Telegram bot.",
    "title": "tgram-analytics",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pageview": {
      "post": {
        "description": "Ingest a pageview event.\n\nForces ``event_name = \"pageview\"`` and stores url/referrer in dedicated\ncolumns as well as in ``properties`` for easy querying.",
        "operationId": "pageview_api_v1_pageview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Pageview Api V1 Pageview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pageview",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/taps": {
      "post": {
        "description": "Ingest the taps and scroll depth of one pageview (tap heatmaps).\n\nSame authentication, rate limits and origin check as /track. Stores one\nrow per tap and one row for the scroll depth. ``session_id`` is not\nstored, no visitor hash is computed, and no alert is evaluated.",
        "operationId": "taps_api_v1_taps_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TapsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Taps Api V1 Taps Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Taps",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/track": {
      "post": {
        "description": "Ingest a custom event.\n\nReturns 202 immediately; alert evaluation runs as a background task.",
        "operationId": "track_api_v1_track_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Track Api V1 Track Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}